pyproject.toml
[project]
name = "qcproject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=6.29.5",
    "latexmk>=0.0.2",
    "matplotlib>=3.10.3",
    "networkx>=3.5",
    "pdflatex>=0.1.3",
    "pydantic-numpy>=8.0.1",
    "pylatexenc>=2.10",
    "qiskit-aer>=0.17.1",
    "qiskit-mps-initializer>=0.2.4",
    "qutip>=5.2.0",
]
